Tidy up DBus helper classes make sure metadata is up-tp-date when switching players (WIP)
